Move endpoint tests to 'endpoint' module
