+++ integration-test/src/test/java/org/jboss/pnc/integration_new/endpoint/GroupConfigEndpointTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.integration_new;
+package org.jboss.pnc.integration_new.endpoint;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
